requirement.txt
pygments
xmltree
numpy
pycairo 
PyGObject
pandas
urllib3
pillow
matplotlib
cairosvg
